master_duel.py: Added betting functionality ALMOST MVP LEVEL
